A bunch of minor changes relating to Elman netorks.   Fixed an error in the distance function in Simbrianmath.   Providing access to all of an srn's layers.   Some name changes in text world.

+++ src/org/simbrain/util/math/SimbrainMath.java
@@ -71,7 +71,7 @@ public class SimbrainMath {
         double y1 = src[1];
         double y2 = tar[1];
 
-        return (int) Math.sqrt(Math.pow((x1 - x2), 2) + Math.pow((y1 - y2), 2));
+        return (double) Math.sqrt(Math.pow((x1 - x2), 2) + Math.pow((y1 - y2), 2));
     }
 
     /**
